Test State
it's gate 2. We have other test stages that would work out better, but gate 2 was the thing I was working on before we decided to start from scratch.
also includes a start to a snowstate implementation. I created a check function so you can seperate the state checks from the state code